+++ TentakelsAttacking2/Helper/public/HTextProcessing.h
