+++ TentakelsAttacking2/UI/Elements/Default/public/InputLine.hpp
@@ -7,7 +7,7 @@
 #include "UIElement.hpp"
 #include "Focusable.h"
 #include "AppContext.h"
-#include "TextProcessing.h"
+#include "HTextProcessing.h"
 #include "HTextProcecingConstant.hpp"
 #include <raylib.h>
 #include <string>
